added more test cases

+++ backend/routes/product.js
@@ -208,13 +208,17 @@ productRouter.get("/products/:id", isLoggedIn, async (req, res) => {
   const isValid = mongoose.isValidObjectId(id);
   console.log(isValid);
   if (!isValid) {
-    return res.sendStatus(StatusCodes.BAD_REQUEST);
+    return res.status(StatusCodes.BAD_REQUEST).send("Invalid Product ID");
   }
 
   try {
     const product = await getProductById(id);
 
     console.log(product);
+
+    if (!product) {
+      return res.status(StatusCodes.NOT_FOUND).send("Product Not Found");
+    }
 
     return res.status(StatusCodes.OK).json(product);
   } catch (error) {
@@ -229,6 +233,10 @@ productRouter.delete("/products/:productId", isLoggedIn, async (req, res) => {
     const productId = req.params.productId;
 
     const deletedProduct = await deleteProduct(productId);
+
+    if (!deletedProduct) {
+      return res.status(StatusCodes.NOT_FOUND).send("Product Not Found");
+    }
 
     // delete all reviews for product
     await ProductReview.deleteMany({ product: productId });
@@ -265,8 +273,6 @@ productRouter.delete("/products/:productId", isLoggedIn, async (req, res) => {
       return res
         .status(StatusCodes.OK)
         .json({ message: "Product deleted successfully" });
-    } else {
-      return res.status(StatusCodes.NOT_FOUND).send("Product not found");
     }
   } catch (error) {
     console.log(error);
